[clang-tidy] Remove redundant cast expression
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/213178
Tested-by: Build Bot <build@couchbase.com>

+++ cluster_framework/node.cc
@@ -111,7 +111,7 @@ NodeImpl::NodeImpl(std::filesystem::path directory, std::string id)
                 {"minidump_dir", minidump_dir.generic_string()},
                 {"content", "default"}}},
               {"portnumber_file", portnumber_file.generic_string()},
-              {"parent_identifier", (int)getpid()}};
+              {"parent_identifier", getpid()}};
     std::ofstream out(configfile.generic_string());
     out << config.dump(2);
     out.close();
