+++ executor/notifiable_task.cc
@@ -21,6 +21,6 @@ bool NotifiableTask::run() {
 void NotifiableTask::wakeup() {
     bool expected = false;
     if (pendingRun.compare_exchange_strong(expected, true)) {
-        ExecutorPool::get()->wake(static_cast<size_t>(uid));
+        ExecutorPool::get()->wake(uid);
     }
 }
